Extract WakeEvent into jfn-wake-event crate and use it everywhere
