+++ src/jfn_rust/Cargo.toml
@@ -31,6 +31,7 @@ jfn-logging = { path = "../logging" }
 jfn-mpv = { path = "../mpv" }
 jfn-paths = { path = "../paths" }
 jfn-playback = { path = "../playback" }
+jfn-wake-event = { path = "../wake_event" }
 
 [target.'cfg(unix)'.dependencies]
 libc.workspace = true
