+++ src/playback/Cargo.toml
@@ -11,16 +11,8 @@ crate-type = ["rlib"]
 [dependencies]
 parking_lot.workspace = true
 jfn-mpv = { path = "../mpv" }
+jfn-wake-event = { path = "../wake_event" }
 serde_json.workspace = true
-
-[target.'cfg(unix)'.dependencies]
-libc.workspace = true
-
-[target.'cfg(windows)'.dependencies]
-windows-sys = { workspace = true, features = [
-    "Win32_Foundation",
-    "Win32_System_Threading",
-] }
 
 [lints]
 workspace = true
